move LanguageSettings UI to settings/language package
Relocates all LanguageSettings source files from `ui/languages` to
`ui/settings/language`, updating package declarations and imports
across all consumers.
