+++ app/src/main/kotlin/org/cru/godtools/ui/settings/language/LanguageSettingsModule.kt
@@ -1,4 +1,4 @@
-package org.cru.godtools.ui.languages
+package org.cru.godtools.ui.settings.language
 
 import dagger.Module
 import dagger.Provides
