+++ app/src/test/snapshots/images/org.cru.godtools.ui.settings.language_LanguageSettingsLayoutPaparazziTest_LanguageSettingsLayout()[Nexus_5,NOTNIGHT,ACCESSIBILITY].png
