+++ app/src/test/snapshots/images/org.cru.godtools.ui.settings.language_LanguageSettingsLayoutPaparazziTest_LanguageSettingsLayout()[Pixel_6_Pro,NOTNIGHT,NO_ACCESSIBILITY].png
