+++ app/src/testDebug/kotlin/org/cru/godtools/ui/settings/language/LanguageSettingsLayoutPaparazziTest.kt
@@ -1,4 +1,4 @@
-package org.cru.godtools.ui.languages
+package org.cru.godtools.ui.settings.language
 
 import app.cash.paparazzi.DeviceConfig
 import com.android.resources.NightMode
@@ -6,10 +6,9 @@ import com.google.testing.junit.testparameterinjector.TestParameter
 import com.google.testing.junit.testparameterinjector.TestParameterInjector
 import java.util.Locale
 import kotlin.test.Test
-import kotlinx.collections.immutable.persistentListOf
 import org.cru.godtools.base.ui.BasePaparazziTest
 import org.cru.godtools.model.Language
-import org.cru.godtools.ui.languages.LanguageSettingsPresenter.UiState
+import org.cru.godtools.ui.settings.language.LanguageSettingsPresenter.UiState
 import org.junit.runner.RunWith
 
 @RunWith(TestParameterInjector::class)
@@ -21,7 +20,7 @@ class LanguageSettingsLayoutPaparazziTest(
     private val state = UiState(
         appLanguage = Locale.FRENCH,
         appLanguages = 1234,
-        downloadedLanguages = persistentListOf(
+        downloadedLanguages = listOf(
             Language(Locale.ENGLISH),
             Language(Locale.FRENCH),
             Language(Locale.GERMAN),
